Prevented extraneous calls to re_Reload__PossiblyAfterRe_Presenting(:) ⚔️ (This hard-to-concisely-name method is a 'nuclear option' bug killer for layout bugs (cells not displaying re-sized in-time) that occur upon device rotations.
The method is (or was) an extension of every collection view controller ('VC') in this app. It gets called from the prepare() method of a VC's layout object.  All VCs in this app have the same class of layout object (a CustomFlowLayout) and thus all call the same prepare().  But— inside prepare(), the method re_Reload…(:) is actually called on the global VC variable topVC, which is the current navigation-controller-embedded VC being displayed (popup-window-style VCs, with embeddedInNavController false, are never the topVC).

So, if any number (call it N) of popup-style VCs were added as subviews of topVC, and, e.g., the device were rotated, then topVC.re_Reload…(:) would be called N+1 times per call to prepare(). This actually caused complete UI failures, where all the collection views, navigation items and bar buttons disappeared, making it impossible to interact with the app*, worse than a crash-and-quit.) This commit fixes these multiple calls with a simple conditional (as there will only ever be 1 topVC, and 1 navigation-controller-embedded VC displayed at once-- thus only 1 call to re_Reload…(:) per call to prepare().)

*This app actually has an invisible navigation bar button, to the left of the leftmost visible one, that simply calls collectionView.reloadData() on topVC, as a sort of emergency UI fix in case it ever still needed updating. Somewhat like a web browser’s refresh button. (It should ideally never be needed though, obviously, and hasn’t: not since probably the predecessor of the predecessor of this project.)
But since the UI failure mentioned above removed all the navigation buttons, this was the one time during this app’s (and its predecessors') development that the button was not able to restore the UI when ‘cells-not-resized-in-time’ type bugs popped up.